+++ src/features/embalm/stepContent/components/ArchaeologistList.tsx
@@ -1,10 +1,8 @@
 import {
   Flex,
-  Image,
   Table,
   TableContainer,
   Tbody,
-  Td,
   Text,
   Th,
   Thead,
@@ -12,7 +10,6 @@ import {
   Button,
   VStack,
   Input,
-  Checkbox,
   HStack,
   Icon,
   Popover,
@@ -26,135 +23,13 @@ import { DownIcon } from 'components/icons/DownIcon';
 import { UpDownIcon } from 'components/icons/UpDownIcon';
 import { UpIcon } from 'components/icons/UpIcon';
 import { Loading } from 'components/Loading';
-import { formatAddress } from 'lib/utils/helpers';
 import { useArchaeologistList } from '../hooks/useArchaeologistList';
-import { SortDirection, selectArchaeologist, deselectArchaeologist } from 'store/embalm/actions';
+import { SortDirection } from 'store/embalm/actions';
 import { FilterInput } from './FilterInput';
-import { ethers } from 'ethers';
-import { useState, Dispatch, SetStateAction, ReactChild } from 'react';
-import { useDialArchaeologists } from '../../../../hooks/utils/useDialArchaeologists';
+import { useState } from 'react';
+// import { useDialArchaeologists } from '../../../../hooks/utils/useDialArchaeologists';
 import { useBootLibp2pNode } from '../../../../hooks/libp2p/useBootLibp2pNode';
-import { useDispatch } from 'store/index';
-
-interface ArchaeologistListItemProps {
-  archaeologist: Archaeologist;
-  isSelected: boolean;
-  includeDialButton: boolean;
-  isDialing: boolean;
-  setIsDialing: Dispatch<SetStateAction<boolean>>;
-  onClick: () => void;
-}
-
-interface TableContentProps {
-  children: React.ReactNode;
-  icon: boolean;
-  checkbox: boolean;
-}
-
-function ArchaeologistListItem({
-  isSelected,
-  archaeologist,
-  includeDialButton,
-  isDialing,
-  setIsDialing,
-  onClick,
-}: ArchaeologistListItemProps) {
-  const [isPinging, setIsPinging] = useState(false);
-  const { testDialArchaeologist } = useDialArchaeologists(setIsDialing);
-  const dispatch = useDispatch();
-
-  function TableContent({ children, icon, checkbox }: TableContentProps) {
-    return (
-      <Td isNumeric>
-        <Flex justify={icon || checkbox ? 'left' : 'center'}>
-          {icon && (
-            <Image
-              src="sarco-token-icon.png"
-              w="18px"
-              h="18px"
-            />
-          )}
-          {checkbox && (
-            <Checkbox
-              isChecked={isSelected && true}
-              onChange={() => {
-                if (isSelected === true) {
-                  dispatch(selectArchaeologist(archaeologist));
-                } else {
-                  dispatch(deselectArchaeologist(archaeologist.profile.archAddress));
-                }
-              }}
-              colorScheme="blue"
-            ></Checkbox>
-          )}
-          <Text
-            ml={3}
-            bg={'brand.100'}
-            py={0.5}
-            px={2}
-          >
-            {children}
-          </Text>
-        </Flex>
-      </Td>
-    );
-  }
-
-  return (
-    <Tr
-      background={isSelected ? 'brand.50' : ''}
-      onClick={() => {
-        onClick();
-        if (!isSelected) {
-          setIsPinging(true);
-        }
-      }}
-      cursor="pointer"
-      _hover={isSelected ? {} : { background: 'brand.0' }}
-    >
-      <TableContent
-        icon={false}
-        checkbox={true}
-      >
-        {formatAddress(archaeologist.profile.archAddress)}
-      </TableContent>
-      <TableContent
-        icon={true}
-        checkbox={false}
-      >
-        {Number(ethers.utils.formatEther(archaeologist.profile.minimumDiggingFee))
-          .toFixed(0)
-          .toString()
-          .concat(' SARCO')}
-      </TableContent>
-      <TableContent
-        icon={false}
-        checkbox={false}
-      >
-        {archaeologist.profile.successes.toString()}
-      </TableContent>
-      <TableContent
-        icon={false}
-        checkbox={false}
-      >
-        {archaeologist.profile.cleanups.toString()}
-      </TableContent>
-
-      {includeDialButton ? (
-        <Td>
-          <Button
-            disabled={isDialing || !!archaeologist.connection}
-            onClick={() => testDialArchaeologist(archaeologist.fullPeerId!)}
-          >
-            {archaeologist.connection ? 'Connected' : 'Dial'}
-          </Button>
-        </Td>
-      ) : (
-        <></>
-      )}
-    </Tr>
-  );
-}
+import { ArchaeologistListItem } from './ArchaeologistListItem';
 
 export function ArchaeologistList({
   includeDialButton,
@@ -191,7 +66,7 @@ export function ArchaeologistList({
   // Used for testing archaeologist connection
   // TODO -- can be removed once we resolve connection issues
   const [isDialing, setIsDialing] = useState(false);
-  const { testDialArchaeologist } = useDialArchaeologists(setIsDialing);
+  // const { testDialArchaeologist } = useDialArchaeologists(setIsDialing);
   useBootLibp2pNode();
 
   return (
